+++ src/main/java/com/example/assignment2/repository/BookingRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 
 
-public interface BookingRepository extends JpaRepository<Booking, Integer>{
+public interface BookingRepository extends JpaRepository<Booking, Long>{
 
-    Optional<Booking> findById(Integer id);
+    Optional<Booking> findById(Long id);
 }
