+++ src/main/java/com/example/assignment2/repository/CustomerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer >{
-    Optional<Customer> findById(Integer id);
+public interface CustomerRepository extends JpaRepository<Customer, Long >{
+    Optional<Customer> findById(Long id);
 }
