+++ src/main/java/com/example/assignment2/repository/InvoiceRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 import java.util.Optional;
-public interface InvoiceRepository extends JpaRepository<Invoice, Integer>{
-    Optional<Invoice> findById(Integer id);
+public interface InvoiceRepository extends JpaRepository<Invoice, Long>{
+    Optional<Invoice> findById(Long id);
 }
